+++ include/loradriver/radio_event.hpp
@@ -4,8 +4,16 @@ namespace loradriver {
 
 enum class RadioEvent {
   kNone = 0,
+  kInitValidate,
+  kInitBindAdapters,
+  kInitHardwareBringUp,
   kInitialized,
-  kTxDone,
+  kTxPreparing,
+  kTxInProgress,
+  kTxCompleted,
+  kTxFailed,
+  kRxListening,
+  kRxInProgress,
   kRxDone,
   kTimeout,
   kError